Serde support
Serde[1] is a serialization framework that converts Rust
data structures to and from a variety of formats: JSON,
YAML, TOML, XML, MessagePack, Bincode.

This commit adds two impls that allow Serde to serialize and
deserialize a LinkedHashMap with any data format supported
by Serde. I have put the impls behind a feature to avoid an
unwanted dependency for people not using Serde.

LinkedHashMap will be valuable to users who need to read a
JSON document and preserve the order of keys in a map, or
write a JSON document and control the order of map keys in
the output.

[1]: https://github.com/serde-rs/serde